Working on Batch Creation for quotes.
WIP: Working on batch creation for quotes. I am currently cloning
everything so I can focus on the logic. I am also not making the
increase, decrease, or stay the same decision for quote price. Also, the
interval has been left out for the moment. I have added TO DOs and
comments that reflect this.